Add files via upload
Javascript Practice. an E-Commerce Cart.
This Project was previously prepared to build the Javascript File Only.you can add products to the cart, if you add more than 1 of the same project, then the quantity will change. and also if you want to remove some product, if there is more than 1 of the same the quantity will change.
you can also empty the cart.
I will like to improve this project creating a floating number which shows trhe number of articles in the cart. 
Another thing will be making the research input work